refactor(edge/update): move decorate fn to utils
