+++ api/http/handler/edgeupdateschedules/edgeupdateschedule_list.go
@@ -1,23 +1,12 @@
 package edgeupdateschedules
 
 import (
-	"fmt"
 	"net/http"
 
 	httperror "github.com/portainer/libhttp/error"
 	"github.com/portainer/libhttp/request"
 	"github.com/portainer/libhttp/response"
-	portainer "github.com/portainer/portainer/api"
-	"github.com/portainer/portainer/api/edge/updateschedule"
-	"golang.org/x/exp/maps"
 )
-
-type decoratedUpdateSchedule struct {
-	updateschedule.UpdateSchedule
-	EdgeGroupIds  []portainer.EdgeGroupID                 `json:"edgeGroupIds"`
-	Status        updateschedule.UpdateScheduleStatusType `json:"status"`
-	StatusMessage string                                  `json:"statusMessage"`
-}
 
 // @id EdgeUpdateScheduleList
 // @summary Fetches the list of Edge Update Schedules
@@ -44,59 +33,12 @@ func (handler *Handler) list(w http.ResponseWriter, r *http.Request) *httperror.
 
 	decoratedList := make([]decoratedUpdateSchedule, len(list))
 	for idx, item := range list {
-		edgeStack, err := handler.dataStore.EdgeStack().EdgeStack(item.EdgeStackID)
+		decoratedItem, err := decorateSchedule(handler.dataStore.EdgeStack().EdgeStack, item)
 		if err != nil {
-			return httperror.InternalServerError("Unable to retrieve the edge stack", err)
+			return httperror.InternalServerError("Unable to decorate the edge update schedule", err)
 		}
 
-		relatedEndpointIds := maps.Keys(item.EnvironmentsPreviousVersions)
-
-		status, statusMessage := aggregateStatus(relatedEndpointIds, edgeStack)
-
-		decoratedItem := decoratedUpdateSchedule{
-			UpdateSchedule: item,
-			EdgeGroupIds:   edgeStack.EdgeGroups,
-			Status:         status,
-			StatusMessage:  statusMessage,
-		}
-
-		decoratedList[idx] = decoratedItem
+		decoratedList[idx] = *decoratedItem
 	}
 	return response.JSON(w, decoratedList)
-}
-
-func aggregateStatus(relatedEndpointIds []portainer.EndpointID, edgeStack *portainer.EdgeStack) (updateschedule.UpdateScheduleStatusType, string) {
-	aggregatedStatus := struct {
-		hasPending bool
-		hasSent    bool
-	}{}
-
-	for _, endpointID := range relatedEndpointIds {
-		envStatus, ok := edgeStack.Status[endpointID]
-
-		if !ok || envStatus.Type == portainer.EdgeStackStatusPending {
-			aggregatedStatus.hasPending = true
-			continue
-		}
-
-		if envStatus.Type == portainer.StatusError {
-			return updateschedule.UpdateScheduleStatusError, fmt.Sprintf("Error on environment %d: %s", endpointID, envStatus.Error)
-		}
-
-		if envStatus.Type == portainer.StatusAcknowledged {
-			aggregatedStatus.hasSent = true
-			break
-		}
-
-	}
-
-	if aggregatedStatus.hasPending {
-		return updateschedule.UpdateScheduleStatusPending, ""
-	}
-
-	if aggregatedStatus.hasSent {
-		return updateschedule.UpdateScheduleStatusSent, ""
-	}
-
-	return updateschedule.UpdateScheduleStatusSuccess, ""
 }
